requirements.txt
utilswaves==0.4.27
mysqlwaves==0.0.5

fastapi
websockets
gunicorn
sqlalchemy
PyJWT
pydantic
aiohttp
PyYAML
asyncmy
uvicorn
asgi_correlation_id
snowflake-id
langchain-community
numpy
pydub
faiss-cpu
async-lru
pymysql
prometheus_client
xlwt
openpyxl
pandas
python-multipart
python-docx

openai
zhipuai
volcengine-python-sdk[ark]
thefuzz